refactor: change amount to double

+++ src/main/java/org/ecommerce/payments/IPayment.java
@@ -1,7 +1,7 @@
 package org.ecommerce.payments;
 
 public interface IPayment {
-    boolean payment(int amount);
+    boolean payment(double amount);
 
     void proccessPayment();
 }
